Load icons from custom directory.
Currently if your icons are located in directory other that `blade-fontawesome`, Factory will throw an error complaining that "path" for a set does not exist. This happens because BladeUIKit package creates "paths" for sets the first time it create Factory. Since this package hooks up to BladeUIKit when its Factory is already instantiated "paths" key for icon sets is not filled.

Ideally, I'd look to update BladeUIKit Factory itself, but it won't hurt to manually sets paths in this library.